more updates in javadocs and site documentation, also misc improvements in common hssf-xssf interfaces
git-svn-id: https://svn.apache.org/repos/asf/poi/trunk@714128 13f79535-47bb-0310-9956-ffa450edef68

+++ src/java/org/apache/poi/hssf/usermodel/HSSFSheet.java
@@ -1656,7 +1656,7 @@ public class HSSFSheet implements org.apache.poi.ss.usermodel.Sheet
      * @param column the column index
      * @param style the style to set
      */
-    public void setDefaultColumnStyle(short column, CellStyle style) {
+    public void setDefaultColumnStyle(int column, CellStyle style) {
         sheet.setDefaultColumnStyle(column, ((HSSFCellStyle)style).getIndex());
     }
 
@@ -1669,7 +1669,7 @@ public class HSSFSheet implements org.apache.poi.ss.usermodel.Sheet
      *
      * @param column the column index
      */
-    public void autoSizeColumn(short column) {
+    public void autoSizeColumn(int column) {
         autoSizeColumn(column, false);
     }
     
@@ -1686,7 +1686,7 @@ public class HSSFSheet implements org.apache.poi.ss.usermodel.Sheet
      * @param column the column index
      * @param useMergedCells whether to use the contents of merged cells when calculating the width of the column
      */
-    public void autoSizeColumn(short column, boolean useMergedCells) {
+    public void autoSizeColumn(int column, boolean useMergedCells) {
         AttributedString str;
         TextLayout layout;
         /**
